# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for what went wrong",
        "title": "What Went Wrong API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/excuses/{id}": {
            "delete": {
                "tags": [
                    "excuses"
                ],
                "summary": "Delete an excuse",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Excuse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid Excuse ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseDeleteErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "excuses"
                ],
                "summary": "Update an excuse",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Excuse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateExcuseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseValidationErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseForbiddenResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseUpdateErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/goals/{goal_id}/excuses": {
            "get": {
                "description": "List excuses, filters by retention days if strictly limited by entitlement.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "excuses"
                ],
                "summary": "List excuses for a goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "goal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From Date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To Date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetExcusesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Goal ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseFetchErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Upsert excuse for a date. Checks entitlement if using premium template.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "excuses"
                ],
                "summary": "Create or update an excuse",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "goal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Excuse Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateExcuseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseValidationErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseForbiddenResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseCreateErrorResponse"
                        }
                    }
                }
            }
        },
        "/goals/{goal_id}/excuses/today": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "excuses"
                ],
                "summary": "Get today's excuse for a goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "goal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Goal ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseFetchErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateExcuseRequest": {
            "type": "object",
            "required": [
                "date",
                "excuseText"
            ],
            "properties": {
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string",
                    "example": "2023-10-27"
                },
                "excuseText": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "寝坊しました。"
                },
                "templateId": {
                    "type": "string",
                    "example": "template_123"
                }
            }
        },
...
        "handlers.ExcuseCreateErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "言い訳の作成に失敗しました"
                }
            }
        },
        "handlers.ExcuseDeleteErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "言い訳の削除に失敗しました"
                }
            }
        },
        "handlers.ExcuseFetchErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "言い訳の取得に失敗しました"
                }
            }
        },
        "handlers.ExcuseForbiddenResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "プレミアムテンプレートを利用するにはプレミアムプランが必要です"
                }
            }
        },
        "handlers.ExcuseNotFoundResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "言い訳が見つかりません"
                }
            }
        },
        "handlers.ExcuseResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string",
                    "example": "2023-10-27"
                },
                "excuseText": {
                    "type": "string",
                    "example": "寝坊しました。"
                },
                "goalId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "templateId": {
                    "type": "string",
                    "example": "template_123"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ExcuseUpdateErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "言い訳の更新に失敗しました"
                }
            }
        },
        "handlers.ExcuseValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "入力内容が正しくありません"
                }
            }
        },
...
        "handlers.GetExcusesResponse": {
            "type": "object",
            "properties": {
                "excuses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ExcuseResponse"
                    }
                }
            }
        },
...
        "handlers.UpdateExcuseRequest": {
            "type": "object",
            "properties": {
                "excuseText": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "盛大に寝坊しました。"
                },
                "templateId": {
                    "type": "string",
                    "example": "template_123"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}